Move acorn code to separate file
